+++ simple_romp/evaluation/eval_Relative_Human.py
@@ -33,7 +33,7 @@ relative_age_types = ['adult', 'teen', 'kid', 'baby']
 relative_depth_types = ['eq', 'cd', 'fd']
 
 def joint_mapping(source_format, target_format):
-    mapping = np.ones(len(target_format),dtype=np.int)*-1
+    mapping = np.ones(len(target_format),dtype=np.int32)*-1
     for joint_name in target_format:
         if joint_name in source_format:
             mapping[target_format[joint_name]] = source_format[joint_name]
